Adds deck reset + try catch fix
Try catch wasn't working if the int wasn't one of the specified, or if
it was a ValueError. Also, if the player was enjoying our game too much
and kept wanting to play over and over the deck would eventually run out
of cards to pop. Thanks, players for enjoying the game that much!